requirements.txt
CairoSVG==2.5.1
Pillow==5.4.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
RPi.GPIO==0.7.0
spidev==3.4
underground==0.3.0
xmljson==0.2.1
